Finalise getAllContainerRepositories and getSubAssessments (#2)
* Initial npm release

* Bump version to 0.0.3

* Update readme

* Update readme

* Begin build of updateAssessmentDatabase function

* Updates for v0.0.3

* Moved to @jercle/yargonaut@1.1.5 to mitigate vulnerability

* Minor updates

* Finalised getAllContainerRepositories func

* Update some options, work on getSubAssessments

* Finalise getSubAssessments

* Working on CLI

* Update args

* Update submodule HEAD
