+++ cli.js
@@ -1,13 +1,18 @@
 const os = require("os")
-const yargs = require("yargs")
 const { DefaultAzureCredential } = require("@azure/identity")
 const clipboardy = require("clipboardy")
 const inquirer = require("inquirer")
-const subList =
-  require(`${os.homedir()}/.azure/azureProfile.json`).subscriptions
+let subList
+try {
+  subList = require(`${os.homedir()}/.azure/azureProfile.json`).subscriptions
+} catch (error) {
+  console.error("Azure CLI profile doesn't seem to be set up")
+  throw new Error(error)
+}
 
 const listShareFiles = require("./funcs/az/listShareFiles")
 const getSecrets = require("./funcs/az/getSecrets")
+const getSubAssessments = require("./funcs/az/getSubAssessments")
 const getAppServicePlan = require("./funcs/az/getAppServicePlan")
 const listSubscriptions = require("./funcs/az/listSubscriptions")
 const getAppService = require("./funcs/az/getAppService")
@@ -34,139 +39,16 @@ require("yargonaut")
   .helpStyle("green.underline")
   .errorsStyle("red.bold")
 
-const options = yargs
-  .check((argv, options) => {
-    if (
-      argv._[0] == "ls" ||
-      argv._[0] == "secrets" ||
-      argv._[0] == "asp" ||
-      argv._[0] == "app" ||
-      argv._[0] == "init" ||
-      argv._[0] == "subs"
-    ) {
-      return true // tell Yargs that the arguments passed the check
-    } else {
-      throw new Error("See above for available commands")
-    }
-  })
-  .usage(`azgo ${"<command>"} ${"[option]"}`)
-  // .usage("ls -n <File Share name> -c <connection string>")
-  // .usage("secrets -a <App Name> -e <environment>")
-  .command("init", "Initiate app, and configuration", (yargs) => {
-    return yargs
-      .option("clearCache", {
-        alias: "x",
-        type: "boolean",
-        description: "Clears CLI cache",
-      })
-      .option("showCache", {
-        alias: "s",
-        type: "boolean",
-        description: "Prints current CLI cache to terminal",
-      })
-  })
-  .command("ls", "List files in given File Share", (yargs) => {
-    return yargs
-      .option("connStr", {
-        describe: "Connection String for Storage Account",
-        alias: "c",
-        type: "string",
-        demandOption: true,
-      })
-      .option("shareName", {
-        describe: "Name of File Share",
-        alias: "n",
-        type: "string",
-        demandOption: true,
-      })
-      .option("path", {
-        describe: "Directory path within File Share",
-        alias: "p",
-        type: "string",
-        default: "",
-      })
-  })
-  .command("secrets", "List secrets for an app's keyvault", (yargs) => {
-    return yargs
-      .option("appName", {
-        describe: "Name of App",
-        alias: "n",
-        type: "string",
-        demandOption: true,
-      })
-      .option("appEnv", {
-        describe: "Environment ('dev', 'test', 'prod'",
-        alias: "e",
-        type: "string",
-        demandOption: true,
-      })
-  })
-  .command("asp", "Get details of an App Service Plan", (yargs) => {
-    return yargs
-      .option("appName", {
-        describe: "Name of App",
-        alias: "n",
-        type: "string",
-        demandOption: true,
-      })
-      .option("appEnv", {
-        describe: "Environment ('dev', 'test', 'prod')",
-        alias: "e",
-        type: "string",
-        demandOption: true,
-      })
-      .option("subscriptionId", {
-        describe: "Subscription ID",
-        alias: "s",
-        type: "string",
-        demandOption: true,
-      })
-  })
-  .command("app", "Get details of an App Service", (yargs) => {
-    return yargs
-      .option("appName", {
-        describe: "Name of App",
-        alias: "n",
-        type: "string",
-        demandOption: true,
-      })
-      .option("appEnv", {
-        describe: "Environment ('dev', 'test', 'prod')",
-        alias: "e",
-        type: "string",
-        demandOption: true,
-      })
-      .option("subscriptionId", {
-        describe: "Subscription ID",
-        alias: "s",
-        type: "string",
-        demandOption: true,
-      })
-  })
-  .command("subs", "List subscriptions in current AZ CLI profile", (yargs) => {
-    return yargs.command(
-      "active",
-      "Select current active subscription for this CLI and Azure CLI"
-    )
-  })
-  .help()
-  .alias("h", "help")
-  .alias("v", "version")
-  .option("copyOutput", {
-    describe: "Copy output to clipboard",
-    alias: "c",
-    default: false,
-    type: "boolean",
-  })
-  .epilogue(
-    "For more information, see the readme at https://github.com/stkcat/azure-tooling"
-  ).argv
-
-function actOnCli() {
+function actOnCli(options) {
+  console.log(options)
   switch (options._[0]) {
     case "init":
       options.clearCache && clearCacheConfirm()
       options.showCache && appCache.show()
+      break
+    case "acr":
+      getSubAssessments(options, azCliCredential)
+      // console.log("acr")
       break
     case "ls":
       listShareFiles(options, azCliCredential)
@@ -206,7 +88,6 @@ function selectSubscription() {
 }
 
 module.exports = {
-  options,
   actOnCli,
   selectSubscription,
 }
