+++ funcs/az/test-getContainerRepository.js
@@ -18,17 +18,31 @@ const { DefaultAzureCredential } = require("@azure/identity")
 //   new DefaultAzureCredential()
 // )
 
-// const opts = {
-//   acrRegistry: "",
-//   subscriptionId: "",
-//   repoName: "gm",
-//   imageRetention: 30,
+const opts = {
+  subscriptionId: process.env.subscriptionId,
+  resourceGroup: process.env.resourceGroup,
+  acrName: process.env.acrName,
+  assessmentName: process.env.assessmentName,
+  nsgName: process.env.nsgName,
+  acrRegistry: process.env.acrRegistry,
+  testDataPath: process.env.testDataPath,
+  appName: process.env.appName
+}
+// {
+//   subscriptionId,
+//   resourceGroup,
+//   acrName,
+//   assessmentName,
+//   nsgName,
+//   acrRegistry,
+//   testDataPath,
 // }
 
-// getContainerRespository(opts)
+
+// getContainerRespository(opts).then(res => console.log(res))
 
 async function getContainerRespository(
-  { repoName, acrRegistry },
+  { appName, acrRegistry },
   azCliCredential
 ) {
   const client = new ContainerRegistryClient(
@@ -41,7 +55,7 @@ async function getContainerRespository(
 
   let manifests = []
 
-  const repo = client.getRepository(repoName)
+  const repo = client.getRepository(appName)
   // console.log(await repo.getProperties())
   for await (manifest of repo.listManifestProperties()) {
     manifests = [...manifests, manifest]
@@ -79,7 +93,7 @@ async function getContainerRespository(
   // }
   // console.log(mans)
 
-  // console.log()
+  // console.log(repository)
   return repository
 }
